Move serial communication to its own singleton class
JetsonPort class will have all functionality required to interface with
Jetson over socket connection.
